Move assert methods from ResourceTest to dedicated utility class #903
This changes moves all assert* utility methods defined in ResourceTest
to a dedicated ResourceTestUtil test class. It prepares for removing the
JUnit 3 test inheritances hierarchy to migrate to JUnit 4.

Contributes to
https://github.com/eclipse-platform/eclipse.platform/issues/903
